fix(config): address config regression introduced by 0.9.2
Revert config_file.rs to its original 0.9.1 and instead of trying to be
smart just duplicate it as config_file/sast.rs and
config_file/secrets.rs so each product has its own copy of the config
parsing implementation.
